requirements.txt
arize-phoenix-otel==0.15.0
fastapi==0.116.1
openai==2.1.0
openinference-instrumentation-openai==0.1.36
opentelemetry-api==1.33.1
pypdf==5.4.0
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.35.0
